Move type ButtonProps to Button component
