+++ components/common/EDS/Button.tsx
@@ -1,8 +1,23 @@
 import { Spinner } from "mad-expo-core";
-import { TouchableOpacity, View } from "react-native";
+import {
+  GestureResponderEvent,
+  StyleProp,
+  TouchableOpacity,
+  View,
+  ViewStyle,
+} from "react-native";
 
-import { ButtonProps } from "../../../types";
 import Typography from "../atoms/Typography";
+
+export type ButtonProps = {
+  onPress: (event: GestureResponderEvent) => void;
+  text: string;
+  outlined?: boolean;
+  danger?: boolean;
+  loading?: boolean;
+  disabled?: boolean;
+  style?: StyleProp<ViewStyle>;
+};
 
 const ButtonEDS = ({
   onPress,
